Add BotState field, refactor schema
Removing all annotations from model makes it more readable. Also fluent
api is more powerful, so all the constraints can be kept in one place
and there is no need after that to look at model to understand what
contstraints are presented there. Model is just for data structure and
Context is for everything else including keys, indexes, relationships
and so on.