Preserve core logic pure
Core logic must be pure, hence easy-testable and implementation- and
infrastructure-agnostic. Returning Async because of only one case that
needs to call database to check if given name is reserved is silly.
Returning Async from Core logic by itself already is not good choice,
because Async is effectfull, hence can not be tested easily.
And returning Async means bad design, which is forced by infrastructure.