Generalize core logic
Using pure synchronous code in core logic is nice, because it's simple
to write, test and understand. But it has a downside. It can't use
asynchronous services. All services have to be synchronous too. This
makes it impossible to use fully asynchronous code, because each service
will block the thread untill is completed. The solution is to use
callbacks! Callbacks are nice, because they generalize the code and
preserve core logic infrastructure-agnostic, still giving the
opportunity to implement it in the asynchronous way. Or any other
way you want.