Change ReportCourseFinished to SaveCourse
SaveCourse sounds better because it gives the intent it real meaning.
Not only report that course is created, but do the main thing wee need
after the course is created — save it. Sending the course data with
SaveCourse intent also looks natural in contrast with
ReportCourseFinished which gives no understanding what should consumer
of the code do with that data. Should it send it to user? Shoud it
report the number of data blocks? No clear definition.